refacotr: Transaction 컴포넌트 로컬 커스텀훅 구현
